Split tests into separate files.
